Split PaletteReference into its own file.
